Separate files with different licenses
